Rework effects logic to use SQLite database.

+++ src/effect/mod.rs
@@ -3,7 +3,6 @@ use std::fs;
 use std::fmt;
 use std::fs::File;
 use std::io::{Read, Write};
-use std::ffi::OsStr;
 use std::path::PathBuf;
 use std::collections::HashMap;
 
@@ -16,31 +15,38 @@ use rand::{thread_rng, Rng};
 use hyper::Client;
 use hyper::header::Connection;
 use edit_distance::edit_distance;
+use flac::{ReadStream, StreamReader, StreamIter};
 
 
 // Modules --------------------------------------------------------------------
 mod effect;
-mod stats;
 
 
 // Internal Dependencies ------------------------------------------------------
 use ::bot::BotConfig;
 use ::server::ServerConfig;
 use ::db::models::{Effect as EffectModel, NewEffect as NewEffectModel};
-//use ::db::schema::effects::dsl::{server_id, name as effect_name};
+use ::db::schema::effects::dsl::{server_id, name as effect_name};
 use ::db::schema::effects::table as effectTable;
-use self::stats::EffectStatCache;
 
-pub use self::stats::EffectStat;
 pub use self::effect::Effect as Effect;
+
+
+// Effect Statistics ----------------------------------------------------------
+#[derive(Debug, Clone)]
+pub struct EffectStat {
+    pub duration_ms: u64,
+    pub peak_db: f32,
+    pub silent_start_samples: u64,
+    pub silent_end_samples: u64
+}
 
 
 // Effects Registration -------------------------------------------------------
 #[derive(Debug)]
 pub struct EffectRegistry {
     effects: HashMap<String, Effect>,
-    last_played: HashMap<String, u64>,
-    stat_cache: EffectStatCache
+    last_played: HashMap<String, u64>
 }
 
 
@@ -50,33 +56,21 @@ impl EffectRegistry {
     pub fn new() -> EffectRegistry {
         EffectRegistry {
             effects: HashMap::new(),
-            last_played: HashMap::new(),
-            stat_cache: EffectStatCache::new()
+            last_played: HashMap::new()
         }
     }
 
     pub fn reload(&mut self, config: &ServerConfig) {
-
-        //use ::db::schema::effects::dsl::server_id;
-
         self.effects.clear();
-
-        //for effect in effectTable.filter(server_id.eq(&self.table_id))
-        //          .load::<EffectModel>(&self.connection)
-        //          .unwrap_or_else(|_| vec![]) {
-
-
-        //}
-
         self.load_effects(config)
     }
 
-    pub fn has_effect(&self, effect_name: &str) -> bool {
-        self.effects.contains_key(effect_name)
+    pub fn has_effect(&self, name: &str) -> bool {
+        self.effects.contains_key(name)
     }
 
-    pub fn get_effect(&self, effect_name: &str) -> Option<&Effect> {
-        self.effects.get(effect_name)
+    pub fn get_effect(&self, name: &str) -> Option<&Effect> {
+        self.effects.get(name)
     }
 
     pub fn played_effect(&mut self, name: &str) {
@@ -148,37 +142,36 @@ impl EffectRegistry {
         &mut self,
         config: &ServerConfig,
         effect: &Effect,
-        effect_name: &str
+        name: &str
 
     ) -> Result<(), String> {
 
         let mut new_effect_path = config.effects_path.clone();
-        let mut new_transcript_path = config.effects_path.clone();
-
         if let Some(uploader) = effect.uploader() {
             new_effect_path.push(format!(
                 "{}.{}.flac",
-                effect_name,
+                name,
                 uploader.replace("#", "_")
             ))
 
         } else {
-            new_effect_path.push(effect_name);
+            new_effect_path.push(name);
             new_effect_path.set_extension("flac");
         }
 
-        new_transcript_path.push(effect_name);
-        new_transcript_path.set_extension("txt");
+        // TODO use a transaction?
+        let q = effectTable.filter(server_id.eq(&config.table_id)).filter(effect_name.eq(effect.name.clone()));
+        if diesel::update(q).set(effect_name.eq(name)).execute(&config.connection).is_ok() {
+            fs::rename(effect.to_path_str(), new_effect_path).map_err(|err| {
+                err.to_string()
 
-        fs::rename(effect.to_path_str(), new_effect_path).map_err(|err| {
-            err.to_string()
+            }).and_then(|_| {
+                Ok(self.reload(config))
+            })
 
-        }).and_then(|_| {
-            // TODO rename effect in DB
-            fs::rename(effect.transcript_path(), new_transcript_path).ok();
-            self.reload(config);
-            Ok(())
-        })
+        } else {
+            Err("Failed to rename effect in database.".to_string())
+        }
 
     }
 
@@ -188,29 +181,33 @@ impl EffectRegistry {
         effect: &Effect
 
     ) -> Result<(), String> {
-        // TODO remove effect to DB
-        fs::remove_file(effect.to_path_str()).map_err(|err| {
-            err.to_string()
+        // TODO use a transaction?
+        let q = effectTable.filter(server_id.eq(&config.table_id)).filter(effect_name.eq(&effect.name));
+        if diesel::delete(q).execute(&config.connection).is_ok() {
+            fs::remove_file(effect.to_path_str()).map_err(|err| {
+                err.to_string()
 
-        }).and_then(|_| {
-            fs::remove_file(effect.transcript_path()).ok();
-            self.reload(config);
-            Ok(())
-        })
+            }).and_then(|_| {
+                Ok(self.reload(config))
+            })
+
+        } else {
+            Err("Failed to delete effect from database.".to_string())
+        }
     }
 
     pub fn download_effect(
         &mut self,
         config: &ServerConfig,
-        effect_name: &str,
+        name: &str,
         upload_url: &str,
         uploader: &str
 
     ) -> Result<(), String> {
-        // TODO add effect to DB and get stats
+
         download_file(
             config.effects_path.clone(),
-            effect_name,
+            name,
             upload_url,
             Some(uploader),
             "flac"
@@ -218,22 +215,60 @@ impl EffectRegistry {
         ).map_err(|err| {
             err.to_string()
 
-        }).and_then(|_| {
-            Ok(self.reload(config))
+        }).and_then(|effect_path| {
+
+            // TODO dry error handling
+            if let Ok(stats) = analyze_flac(&effect_path) {
+
+                let new_effect = NewEffectModel {
+                    server_id: &config.table_id,
+                    name: name,
+                    uploader: uploader,
+                    peak_db: stats.peak_db,
+                    duration_ms: stats.duration_ms as i32,
+                    silent_start_samples: stats.silent_start_samples as i32,
+                    silent_end_samples: stats.silent_end_samples as i32,
+                    transcript: ""
+                };
+
+                if diesel::insert(&new_effect).into(effectTable).execute(&config.connection).is_ok() {
+                    Ok(self.reload(config))
+
+                } else {
+                    fs::remove_file(effect_path).map_err(|err| {
+                        err.to_string()
+
+                    }).and_then(|_| {
+                        Err("Failed to analyze uploaded flac file.".to_string())
+                    })
+                }
+
+            } else {
+                fs::remove_file(effect_path).map_err(|err| {
+                    err.to_string()
+
+                }).and_then(|_| {
+                    Err("Failed to analyze uploaded flac file.".to_string())
+                })
+            }
+
         })
+
     }
 
     pub fn download_transcript(
         &mut self,
-        config: &ServerConfig,
-        effect_name: &str,
-        upload_url: &str
+        _: &ServerConfig,
+        _: &str,
+        _: &str
 
     ) -> Result<(), String> {
         // TODO update transcript for effect in DB
+        // TODO download text instead of file
+        /*
         download_file(
             config.effects_path.clone(),
-            effect_name,
+            name,
             upload_url,
             None,
             "txt"
@@ -244,6 +279,8 @@ impl EffectRegistry {
         }).and_then(|_| {
             Ok(self.reload(config))
         })
+        */
+        Ok(())
     }
 
 }
@@ -342,42 +379,55 @@ impl EffectRegistry {
 
     fn load_effects(&mut self, config: &ServerConfig) {
 
-        // TODO load from database
-
         let start = clock_ticks::precise_time_ms();
-        filter_dir(&config.effects_path, "flac", |name, path| {
+        for effect in effectTable.filter(server_id.eq(&config.table_id))
+                  .load::<EffectModel>(&config.connection)
+                  .unwrap_or_else(|_| vec![]) {
 
-            // let duration = get_flac_duration(path.clone()).unwrap_or(0);
-            let descriptor: Vec<&str> = name.split('.').collect();
-            let effect = if descriptor.len() == 2 {
-                Effect::new(
-                    descriptor[0],
-                    path.clone(),
-                    self.stat_cache.get(config, path, descriptor[0]),
-                    Some(descriptor[1].replace("_", "#"))
-                )
-
-            } else {
-                Effect::new(
-                    descriptor[0],
-                    path.clone(),
-                    self.stat_cache.get(config, path, descriptor[0]),
-                    None
-                )
-            };
-
-            self.effects.insert(
-                effect.name.clone(),
-                effect.with_transcript()
-            );
-
-        });
+            let effect = self.effect_from_model(config, effect);
+            self.effects.insert(effect.name.clone(), effect);
+        }
 
         info!(
             "{} Effects loaded in {}ms.",
             self,
             clock_ticks::precise_time_ms() - start
         );
+
+    }
+
+    fn effect_from_model(
+        &self,
+        config: &ServerConfig,
+        effect: EffectModel
+
+    ) -> Effect {
+
+        let mut path = PathBuf::new();
+        path.push(config.effects_path.clone());
+
+        if effect.uploader.is_empty() {
+            path.push(effect.name.clone());
+
+        } else {
+            path.push(format!("{}.{}.", effect.name, effect.uploader));
+        }
+
+        path.set_extension("flac");
+
+        let e = Effect::new(
+            effect.name.as_str(),
+            path,
+            Some(EffectStat {
+                duration_ms: effect.duration_ms as u64,
+                peak_db: effect.peak_db,
+                silent_start_samples: effect.silent_start_samples as u64,
+                silent_end_samples: effect.silent_end_samples as u64
+            }),
+            Some(effect.uploader)
+        );
+
+        e.with_transcript(effect.transcript)
 
     }
 
@@ -487,46 +537,20 @@ fn match_alias_pattern(alias: &str, pattern: &str) -> bool {
 
 }
 
-fn filter_dir<F: FnMut(String, PathBuf)>(
-    path: &PathBuf,
-    ext: &str,
-    mut callback: F
-) {
-    if let Ok(listing) = fs::read_dir(path) {
-        for entry in listing {
-            if let Ok(entry) = entry {
-                if entry.file_type().unwrap().is_file() {
-                    let path = entry.path();
-                    if path.extension().unwrap_or_else(|| OsStr::new("")) == ext {
-                        if let Some(stem) = path.file_stem() {
-                            if stem != "" {
-                                callback(
-                                    stem.to_str().unwrap_or("").to_string(),
-                                    PathBuf::from(path.clone())
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
 fn download_file(
     mut directory: PathBuf,
-    effect_name: &str,
+    name: &str,
     url: &str,
     nickname: Option<&str>,
     ext: &str
 
-) -> Result<(), String> {
+) -> Result<PathBuf, String> {
 
     if let Some(nickname) = nickname {
-        directory.push(&format!("{}.{}.{}", effect_name, nickname, ext));
+        directory.push(&format!("{}.{}.{}", name, nickname, ext));
 
     } else {
-        directory.push(&format!("{}.{}", effect_name, ext));
+        directory.push(&format!("{}.{}", name, ext));
     }
 
     let client = Client::new();
@@ -541,13 +565,91 @@ fn download_file(
                 .map(|_| buffer)
         })
         .and_then(|buffer| {
-            File::create(directory)
+            File::create(directory.clone())
                 .map_err(|err| err.to_string())
                 .and_then(|mut file| {
                     file.write_all(&buffer)
                         .map_err(|err| err.to_string())
+                        .and_then(|_| Ok(directory))
                 })
         })
 
 }
 
+fn analyze_flac(flac_path: &PathBuf) -> Result<EffectStat, String> {
+    StreamReader::<File>::from_file(flac_path.to_str().unwrap_or(""))
+        .map_err(|_| "Failed to open flac file.".to_string())
+        .and_then(|stream| {
+            Ok(analyze_flac_stream(stream))
+        })
+}
+
+fn analyze_flac_stream(stream: StreamReader<File>) -> EffectStat {
+
+    let stream_info = stream.info();
+    let samples: StreamIter<ReadStream<File>, i64> = StreamIter::new(stream);
+
+    let mut sample_count = 0;
+    let mut last_active_sample = 0;
+
+    let sum_squares = samples.into_iter().fold(0.0f64, |acc, s| {
+        let sample = s as f64 / 32768.0;
+        if sample > 0.01 {
+            sample_count += 1;
+            if sample > 0.025 {
+                last_active_sample = sample_count;
+            }
+            acc + sample.powf(2.0f64)
+
+        } else {
+            acc
+        }
+    });
+
+    let rms = (sum_squares / (sample_count as f64)).sqrt();
+    EffectStat {
+        duration_ms: (stream_info.total_samples * 1000) / stream_info.sample_rate as u64,
+        peak_db: (20.0 * rms.log(10.0)) as f32,
+        silent_start_samples: 0,
+        silent_end_samples: sample_count - last_active_sample
+    }
+
+}
+/*
+fn load_transcript(mut flac_path: PathBuf) -> Option<String> {
+
+    flac_path.set_extension("txt");
+
+    if let Ok(mut file) = File::open(flac_path) {
+
+        let mut text = String::new();
+        file.read_to_string(&mut text).expect("Failed to read flac transcript.");
+
+        // Remove linebreaks
+        text = text.to_lowercase().replace(|c| {
+            match c {
+                '\n' | '\r' | '\t' => true,
+                _ => false
+            }
+
+        }, " ");
+
+        // Split up into unique words
+        let mut parts: Vec<String> = text.split(' ').filter(|s| {
+            !s.trim().is_empty()
+
+        }).map(|s| {
+            s.to_string()
+
+        }).collect();
+
+        parts.dedup();
+
+        Some(parts.join(" "))
+
+    } else {
+        None
+    }
+
+}
+*/
